refactor(test): separate KaonaviClient test
